fix(testing): Fix linting and type errors in testing framework
- Fix ruff linting issues (unused imports, import sorting)
- Fix mypy type errors in TestRegistry, ReportGenerator
- Add type annotations to synthetic_image.py for Python 3.9 compatibility
- Fix compute_dice return type in test files
- Exclude TestCases/ from mypy (depends on Slicer runtime)
- Update pyproject.toml package discovery for multiple modules
- Install pre-commit hooks

+++ SegmentEditorAdaptiveBrush/Testing/Python/test_adaptive_algorithm.py
@@ -51,7 +51,7 @@ def requires_sitk(test_func):
 def compute_dice(mask1: np.ndarray, mask2: np.ndarray) -> float:
     """Compute Dice coefficient between two binary masks."""
     intersection = np.sum((mask1 > 0) & (mask2 > 0))
-    return 2 * intersection / (np.sum(mask1 > 0) + np.sum(mask2 > 0) + 1e-8)
+    return float(2 * intersection / (np.sum(mask1 > 0) + np.sum(mask2 > 0) + 1e-8))
 
 
 class TestConnectedThresholdAlgorithm(unittest.TestCase):
